// src/@types/open-news-types.ts
interface ArticleSource {
  id: string | null
  name: string
}

interface Article {
  source: ArticleSource
  author: string | null
  title: string
  description: string | null
  url: string
  urlToImage: string | null
  publishedAt: string
  content: string | null
}

export interface OpenNewsApiResponse {
  status: string
  totalResults: number
  articles: Article[]
}